backend_python/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
supabase==2.10.0
pydantic==2.10.4
textblob==0.18.0
python-dotenv==1.0.1
gunicorn==23.0.0
scikit-learn==1.6.1
numpy==2.2.1
joblib==1.4.2
httpx<0.28,>=0.26
pytest==8.3.4
